+++ src/main/java/com/syshub/modules/articles/repositories/ArticleRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Long>, JpaSpecificationExecutor<Article> {
+public interface ArticleRepository extends JpaRepository<Article, Integer>, JpaSpecificationExecutor<Article> {
     Optional<Article> findBySlug(String slug);
     boolean existsBySlug(String slug);
 }
