+++ src/main/java/com/syshub/modules/catalog/repositories/CourseRepository.java
@@ -4,11 +4,11 @@ import com.syshub.modules.catalog.entities.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 
-    List<Course> findBySemesterNumero(Long semesterId);
+    List<Course> findBySemesterNumero(Integer semesterId);
 
-    List<Course> findByAreaId(Long areaId);
+    List<Course> findByAreaId(Integer areaId);
 
     List<Course> findByNombreContainingIgnoreCaseOrCodigoContainingIgnoreCase(String nombre, String codigo);
 }
