+++ src/main/java/com/syshub/modules/catalog/repositories/PensumRepository.java
@@ -4,6 +4,6 @@ import com.syshub.modules.catalog.entities.Pensum;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface PensumRepository extends JpaRepository<Pensum, Long> {
+public interface PensumRepository extends JpaRepository<Pensum, Integer> {
     List<Pensum> findByCarreraId(Integer carreraId);
 }
