+++ src/main/java/com/syshub/modules/catalog/repositories/TechnicalAreaRepository.java
@@ -3,5 +3,5 @@ package com.syshub.modules.catalog.repositories;
 import com.syshub.modules.catalog.entities.TechnicalArea;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TechnicalAreaRepository extends JpaRepository<TechnicalArea, Long> {
+public interface TechnicalAreaRepository extends JpaRepository<TechnicalArea, Integer> {
 }
