+++ src/main/java/com/syshub/modules/repo/repositories/ProjectRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import java.util.List;
 
-public interface ProjectRepository extends JpaRepository<Project, Long>, JpaSpecificationExecutor<Project> {
-    List<Project> findByCursoId(Long courseId);
+public interface ProjectRepository extends JpaRepository<Project, Integer>, JpaSpecificationExecutor<Project> {
+    List<Project> findByCursoId(Integer courseId);
     List<Project> findByDestacadoTrue();
 }
